requirements.txt
appdirs
autocorrect
blessings
boto
bpython
bz2file
chainer
chardet
cucco
curtsies
cycler
dask
decorator
docopt
filelock
flake8
gensim
greenlet
h5py
ipython
ipython-genutils
jedi
lxml
Markdown
matplotlib
mccabe
neologdn
networkx
nltk
nose
numpy
olefile
packaging
pandas
pdfminer.six
pdfminer2
pdfminer3k
pdfx
pexpect
pickleshare
Pillow
ply
prompt-toolkit
protobuf
ptpython
ptyprocess
py
pycodestyle
pycrypto
pyflakes
Pygments
pyparsing
PyPDF2
pytest
python-dateutil
pytils
pytz
PyWavelets
PyYAML
refextract
requests
scikit-image
scikit-learn
scipy
simplegeneric
six
smart-open
tensorflow
textblob
tools
tqdm
traitlets
Unidecode
virtualenv
wcwidth
Werkzeug
